[ADD] PLayers List in gameDetails Activity working map data in creating and editing game

+++ android-client-1/src/com/blstream/ctf1/converter/JSONConverter.java
@@ -109,7 +109,7 @@ public class JSONConverter {
 		// JSONObject jsonLatLng = jsonLocalization.getJSONObject("latLng");
 		Localization localization = new Localization();
 		localization.setName(jsonLocalization.getString("name"));
-		localization.setRadius(jsonLocalization.getLong("radius"));
+		localization.setRadius(jsonLocalization.getInt("radius"));
 		result.setLocalization(localization);
 		result.setId(jsonObject.getString("id"));
 		result.setName(jsonObject.getString("name"));
